PR-B-5: extract PatternBinder (per-pattern-kind binding)
Move the per-pattern-kind binders out of Binder.cs into a new
'internal sealed class PatternBinder' at
src/Core/CodeAnalysis/Binding/PatternBinder.cs:

  - BindPattern              (dispatch, was Binder.cs ~line 5637)
  - BindConstantPattern      (was Binder.cs ~line 5658)
  - BindTypePattern          (was Binder.cs ~line 5689)
  - BindPropertyPattern      (was Binder.cs ~line 5701)
  - BindRelationalPattern    (was Binder.cs ~line 5725)
  - BindListPattern          (was Binder.cs ~line 5738)

The discard-pattern case is inlined in the BindPattern dispatch
because it has no per-kind logic beyond constructing a
BoundDiscardPattern; it leaves with the dispatch method rather
than warranting a helper.

PatternBinder consumes BinderContext for the diagnostics bag and
the (mutable) Scope used by BindTypePattern, and
ConversionClassifier for the relational-pattern argument
conversion. The callbacks it needs back into the still-on-Binder
helpers (BindExpression for constant-pattern RHS, BindTypeClause
for type-pattern target type, and the IsNilLiteral nil-literal
test for nullable constant patterns) are injected through narrow
Func seams in the constructor — the same pattern established by
ConversionClassifier in PR-B-3 and OverloadResolver in PR-B-4.
PatternBinder never back-references Binder.

What stays on Binder (deliberately):

  - BindSwitchExpression     — statement/expression-level glue
                                (discriminant binding, arm walking,
                                exhaustiveness reporting via
                                ExhaustivenessAnalyzer, narrowing-frame
                                management, per-arm result conversion).
                                Will land in PR-B-7 StatementBinder /
                                PR-B-9 ExpressionBinder.
  - BindPatternSwitchStatement — same reasoning.
  - ExhaustivenessAnalyzer.cs — already a separate file.
  - IsNilLiteral / TryClassifyPatternNarrowing — used by call sites
                                outside the per-pattern binders (binary
                                expression equality, statement-level
                                narrowing). Injected as a Func into
                                PatternBinder rather than widened.

Binder.cs constructs 'new PatternBinder(binderCtx, conversions,
…)' in its constructor and stores it as 'private readonly
PatternBinder patterns'. The two BindPattern call sites
(BindPatternSwitchStatement and BindSwitchExpression) now route
through patterns.BindPattern(...). No public or internal symbol
visible to Compilation.cs, HoverComputer.cs, or any test was
renamed.

Validation:
  - Build clean (0 warnings, 0 errors).
  - Core.Tests pass: 2120 passed, 1 skipped (RegenerateBaseline).
  - RefactoringBaselineTests pass: 92 passed, zero hash diffs —
    the IL byte-identical gate from PR-0 still holds.
  - Interpreter.Tests pass: 72 / 72.
  - LanguageServer.Tests pass: 242 / 242.
  - Sdk.Tests pass: 24 / 24.

LoC delta:
  - Binder.cs: -110 lines (14690 → 14580).
  - PatternBinder.cs: +238 lines (new file).

Next in the sequence: PR-B-6 LambdaBinder.
